Move system tests file now only used for Publisers
No longer relevant under "other" folder, but "publishers" one.
